4. Container orchestration and serverless/storage-trigger/requirements.txt
google-api-core
grpcio
google-cloud-storage==2.10.0
google-cloud-vision==3.1.3
functions-framework==3.3.0